# packages/shared-schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wenyi API",
    "description": "Web API and background workers for Wenyi's translation engine.",
    "version": "0.1.dev555+gd99d53be2.d20260923"
  },
  "paths": {
...
    "/projects/{pid}/review/run": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Run Ai Review",
        "operationId": "run_ai_review_projects__pid__review_run_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReviewRunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobEnqueued"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{pid}/review/runs": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_projects__pid__review_runs_get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewRun"
                  },
                  "title": "Response List Runs Projects  Pid  Review Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{pid}/review/runs/{rid}": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Get Run",
        "operationId": "get_run_projects__pid__review_runs__rid__get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "rid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{pid}/review/{ci}": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Get Chapter For Review",
        "operationId": "get_chapter_for_review_projects__pid__review__ci__get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ci"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterSegments"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{pid}/review/{ci}/segments/{seg_idx}": {
      "put": {
        "tags": [
          "review"
        ],
        "summary": "Edit Segment",
        "operationId": "edit_segment_projects__pid__review__ci__segments__seg_idx__put",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ci"
            }
          },
          {
            "name": "seg_idx",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Seg Idx"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Edit Segment Projects  Pid  Review  Ci  Segments  Seg Idx  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{pid}/review/{ci}/segments/{seg_idx}/history": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Segment History",
        "operationId": "segment_history_projects__pid__review__ci__segments__seg_idx__history_get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ci"
            }
          },
          {
            "name": "seg_idx",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Seg Idx"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SegmentRevision"
                  },
                  "title": "Response Segment History Projects  Pid  Review  Ci  Segments  Seg Idx  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{pid}/review/{ci}/complete": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Mark Reviewed",
        "operationId": "mark_reviewed_projects__pid__review__ci__complete_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            }
          },
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ci"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Mark Reviewed Projects  Pid  Review  Ci  Complete Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChapterSegments": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "title_translated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Translated"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/SegmentOut"
            },
            "type": "array",
            "title": "Segments"
          },
          "review_issues": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Review Issues",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "index",
          "segments"
        ],
        "title": "ChapterSegments"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobEnqueued": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "project_id",
          "kind"
        ],
        "title": "JobEnqueued"
      },
...
      "ReviewItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "issue",
              "change",
              "publication",
              "soft_finding"
            ],
            "title": "Kind"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": ""
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          },
          "suggestion": {
            "type": "string",
            "title": "Suggestion",
            "default": ""
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "fixed",
              "failed",
              "unchanged"
            ],
            "title": "Status"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewLocation"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/ReviewLocation"
            },
            "type": "array",
            "title": "Evidence"
          },
          "issue": {
            "additionalProperties": true,
            "type": "object",
            "title": "Issue"
          },
          "changes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Changes"
          },
          "publications": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Publications"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status"
        ],
        "title": "ReviewItem"
      },
      "ReviewLocation": {
        "properties": {
          "chapter": {
            "type": "integer",
            "title": "Chapter"
          },
          "text_index": {
            "type": "integer",
            "title": "Text Index"
          },
          "segment_index": {
            "type": "integer",
            "title": "Segment Index"
          },
          "chapter_title": {
            "type": "string",
            "title": "Chapter Title"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "current_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Target"
          }
        },
        "type": "object",
        "required": [
          "chapter",
          "text_index",
          "segment_index",
          "chapter_title",
          "source",
          "current_target"
        ],
        "title": "ReviewLocation"
      },
      "ReviewRun": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "review_id": {
            "type": "string",
            "title": "Review Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "issues": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Issues"
          },
          "soft_findings": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Soft Findings"
          },
          "changes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Changes"
          },
          "autofix": {
            "additionalProperties": true,
            "type": "object",
            "title": "Autofix"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "review_id",
          "status"
        ],
        "title": "ReviewRun"
      },
      "ReviewRunRequest": {
        "properties": {
          "autofix": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Autofix"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ReviewRunRequest"
      },
      "SegmentEdit": {
        "properties": {
          "target": {
            "type": "string",
            "title": "Target"
          },
          "expected_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Target"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target",
          "expected_target"
        ],
        "title": "SegmentEdit"
      },
      "SegmentOut": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "display_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Target"
          },
          "target_before_polish": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Before Polish"
          },
          "anchor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "text"
          }
        },
        "type": "object",
        "required": [
          "index",
          "source"
        ],
        "title": "SegmentOut"
      },
      "SegmentRevision": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "translation",
              "polish",
              "manual",
              "update",
              "snapshot",
              "before_polish"
            ],
            "title": "Kind"
          },
          "before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before"
          },
          "after": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "After"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "before",
          "after",
          "created_at"
        ],
        "title": "SegmentRevision"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}